review

+++ dali/kernels/imgproc/color_manipulation/brightness_contrast.h
@@ -28,7 +28,7 @@
 namespace dali {
 namespace kernels {
 
-template<typename OutputType, typename InputType, size_t ndims = 3>
+template<typename OutputType, typename InputType, int ndims = 3>
 class BrightnessContrastCpu {
  private:
   static constexpr size_t spatial_dims = ndims - 1;
